+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - The following server-side APIs to send data to the client:
   - `Connection::send_blocklist`
+  - `Connection::send_internal_network_list`
   - `Connection::send_tor_exit_node_list`
 
 ## [0.8.1] - 2024-11-15